fix remaining typing problems, move back registry
