+++ tests/next_tests/unit_tests/ffront_tests/test_nd_array_field.py
@@ -96,6 +96,8 @@ def product_nd_array_implementation(request):
 
 def test_mixed_fields(product_nd_array_implementation):
     first_impl, second_impl = product_nd_array_implementation
+    if first_impl.__name__ == "cupy" and second_impl.__name__ == "numpy":
+        pytest.skip("Binary operation between CuPy and NumPy requires explicit conversion.")
 
     inp_a = [-1.0, 4.2, 42]
     inp_b = [2.0, 3.0, -3.0]
